[css2] Drop duplicated dfn of `@media` at-rule
The `@media` at-rule was defined twice, first time in the heading and second
time in the prose without the `@` prefix. Both were flagged as "exported".
This update replaces the second one with a reference to the first to avoid
the duplication (and avoid the creation of an at-rule without the `@` prefix).