[css-contain-2] rename `ContentVisibilityAutoStateChanged` to `ContentVisibilityAutoStateChange`
Per https://github.com/w3c/csswg-drafts/issues/7603#issuecomment-1309133575 the group resolved to use present tense for the event and interface.

Since then, Mozilla and Chrome both implemented the interface under the new name.
https://hg.mozilla.org/mozilla-central/rev/6095dabf5574
https://chromium.googlesource.com/chromium/src/+/9851571ecb54c47a205b2321006577b9799c0ea3

This updates both the event name and interface to reflect the resolution and to match implementation.