[css-overflow-3][css-background-3] Define rounding of clipping region established by 'overflow'.
This formally defines the rounding of the clipping region established by
overflow, which was previously defined (other than for 'overflow:clip')
only in a "For example" in the backgrounds spec.

It writes that definition in order to implement the resolution in #7434
(regarding mixes of 'visible' and 'clip' on different axes), and also to
correctly account for overflow-clip-margin.

Finally, it adjusts the aforementioned example for the resolution
in #7434 as well.

Fixes #7434.