[css-contain-3] Tweak the spec for style() queries #8127
Remove the recently added grammar part for style-feature that was
inaccurate.

Do not accept invalid declarations or unsupported property names as
a style-feature. The effect of that is that we do not need to store
invalid declarations, but merely fall back to parse the style() query
with the invalid declaration as a general-enclosed, which will evaluate
to unknown.