[cssom-view-1] Update spec text for visual viewport scroll and scrollend events (#8205)
Adds normative text for firing the scrollend event on the VisualViewport.

scroll was already specified but incorrectly. The text as written meant that a scroll event would have to be dispatched at VisualViewport whenever the document scrolls, even if the VisualViewport didn't scroll.

This PR splits the perform a scroll steps into one for scrolling boxes (the existing one) and one for viewports (newly added). The viewport version distributes the scroll between the visual and layout viewports and the uses the scrolling box version to perform the scroll of each.

Fixes #8103