Revert "Temporary fix for w3c.github.io redirect breakage"
This reverts the series of commits made to attempt to cause the
unversioned spec URLs to work again. Given that drafts.csswg.org is
now simply proxying w3c.github.io/csswg-drafts, the approach of setting
up the redirects in the way that series of commits did won’t work.