[css-view-transitions-1] Resolve open issues. (#8948)
* [css-view-transitions-1] Resolve open issues

This change fixes the following issues inlined into the spec:

* The link for "paint order" doesn’t seem right. Is there a more canonical definition?

The DOM order of `::view-transition-group` pseudo-elements needs to match the paint order of the corresponding DOM elements. This is why elements are traversed in paint order when adding them to named elements. The definition of paint order used here is correct.

* Isolation and the dynamic setting of blending is only necessary...

This was resolved in https://github.com/w3c/csswg-drafts/issues/7814.

* Refactor this so the algorithm takes a set of elements that will be captured...

This was a problem because there was no way to refer to an element participating in a transition. Presence of view-transition-name was not enough. Now there's a bool on element to refer to this state.

* Specify what happens with mix-blend-mode.

Adds spec text to clarify this.

The issues remaining are related to web animations, will take those up in a follow up PR.

* Remove mix-blend-mode changes