[web-animations-1][editorial] Move “animation frame” section out of the timeline definition
No change to text (except headings).

This section was making it hard to understand the “Timelines” section
because it split the generic dfn from the specific types dfns with
a much more complicated concept relating timelines to animations
and events and other things. This moves it out of the way, and gives
it a higher-level section heading (as it deserves).