[css-syntax-3] Rename `<an+b>` to `<an-b>`
See context in #9473. Rename is aimed at avoiding interpretation of `+` as a
component value multiplier. The ID of the definition does not change.

The rename is also done in the definition of the `serialize an <an+b> value`
algorithm, preserving the previous ID since the algorithm was exported (and
referenced by other specs).

Once construct has been renamed, the following specs will need to be slightly
updated as well (they either reference `<an+b>` or `serialize an <an+b> value`):
- cssom-1
- css-gcpm-4
- selectors-4

To be done separately, once the cross-references database has been updated.