[css-view-transitions-1] Fix tree-scoping from being element-based to being name-based.
The previous fix for using the tree context made it inconsistent with how shadow DOM styling
works for things like anchor positioning, and made it so that e.g. `::part` cannot set a
`view-transition-name`.

Refactored to make the `view-transition-name` a tree-scoped name, rather than check the
tree context of the element.

Closes #10145