[css-view-transitions-2] Initial spec for nested view transitions
This specifies the new 'view-transition-group' CSS property and how it used,
with some overview and an initial example.

Based on [this resolution](https://github.com/w3c/csswg-drafts/issues/10334#issuecomment-2165761156).

There are still a lot of open issues but we will open them separately to avoid inflating the initial issue.

Closes #10334