[css-view-transitions] Rename UpdateCallback to something more specific.
The WebIDL namespace is global, it's a bit better if we call this
something a bit more view-transitions-specific so that other specs don't
rely on it, since UpdateCallback is otherwise very generic.

I don't think this would be a normative change since the name of a
callback isn't exposed anywhere author visible.