[css-values] Total attr()-tainting for substitution values
This may have been the intended meaning of the spec text already,
but I think we should clearly state the outcome of #5092:
partial tainting of custom properties was proposed in that issue,
discussed, objected to, and therefore *not* accepted.

This PR hopefully makes it clear that attr()-tainting applies
to whole substitution values, for now.