[CSS-Pseudo] Correct custom properties in highlights
It was resolved in https://github.com/w3c/csswg-drafts/issues/11317, CSS Highlight Pseudos should allow custom property definitions within the rule itself, but take any missing custom properties from the originating element.

This PR updates the spec to re-list custom properties as supported on highlights. Futhermore, it fixes the description of custom property inheritance that was wrong even for the prior version of the spec.