[web-animations-2] Use `dfn` type for animation trigger values (#11664)
* [web-animations-2] Use `dfn` type for animation trigger values

The values for "animation trigger state" and "animation trigger type" were
defined with a `value` definition type. That type is reserved for CSS values:
https://speced.github.io/bikeshed/#dfn-types

This makes tools that crawl specs to extract machine-readable information
wrongly assume that the spec introduces new CSS values.

There is no specific definition type available for generic values. This update
uses a regular `dfn` and updates references to these values accordingly.

* Link animation trigger type values back to dfns in css-animations-2

Per feedback, the values are indeed CSS values for the `animation-trigger-type`
property defined in CSS Animations 2. This update replaces the definitions in
web-animations-2 by references to the definitions in CSS Animations 2 not to
end up with two places that define the values.