[css-values-5] Pseudo element and tree-scope for sibling functions #9573 (#11830)
Per resolution in #9573, element-backed pseudo elements that refer to an
element in an inner scope should not expose their tree order to tree
counting functions. Spec text now says they instead resolve to 0,
loosely based on the minutes[1].

Additionally, re-word the text for how tree counting functions resolve
against pseudo elements. This is not backed by a resolution, but backed
by the discussions in the meeting minutes[1].

[1] https://github.com/w3c/csswg-drafts/issues/9573#issuecomment-2622696411