[css-color-hdr-1] Use definition of `color()` from css-color-5
Issue #11954 noted that the spec does not really extend the definition of the
`color()` function but rather extends the definition of one of the types that
gets used in the definition of `color()`. A previous update dropped duplicated
constructs accordingly, but the duplication of the definition itself remained.

Duplication is confusing for tools and readers because there's no longer any
direct linnk between the definition of the `color()` function and its syntax.

This replaces the definition with a reference to the actual definition in
`css-color-5`, allowing readers to access the actual function syntax. Note
the "link default" should only be needed on a temporary basis (and it is needed
precisely because the cross-references database currently contains more than
one definition of the `color()` function).