[css-values-5] Fix dfn of `<input-position>` and `<output-value>`
A `<dfn>` of type `type` should only enclose a basic data type (something
between `<` and `>`). At least, that's been the case across specs until now.
The spec was using the dfn to describe a more complex `value`. More likely, the
spec was trying to combine the definition of the two types with the definition
of the value pattern that uses them in one single dfn.

This update changes the value pattern definition type to `value` and creates
additional `type` dfns for the individual types so that they can be referenced
from other constructs.