[css-mixins-1] Let 'inherit' work like 'inherit()' (#12387)
With @function, it becomes possible to have a "type mismatch"
between an "outer" custom property and an "inner" custom property:

  @property --x {
    syntax: "auto | red";
    /* ... */
  }

  @function --f(--x <color>: inherit) {
    result: var(--x);
  }

  div {
    --x: red;
    --f(); /* => ? */
  }

We have to address this somehow, and the obvious way seems to be
making 'inherit' behave like "a var(--x) that resolves in the parent
stack frame". This is essentially what inherit() does, which already
exists in the spec.