[css-values-4] Allow `{A}` and `?` multipliers to be stacked
The Value Definition Syntax defines the list of multipliers that can be
stacked.

The syntax definition of `<cursor-image>` in css-ui-4 has started to use
`<number>{2}?` but the `{A}?` stack wasn't explicitly allowed (and some CSS
parsers do not support it yet).

This update completes the list of multipliers that can be stacked with `{A}?`
and `{A,B}?`.