[css-navigation-1] Add :link-to() pseudo-class (with navigation-param() function).
This adds an initial proposal for the :link-to() pseudo-class and a
navigation-param() function that goes inside of it.

This also refactors some of the invocations of the URL pattern "match"
algorithm to structure the parameters the way the algorithm expects
(and be a little bit clearer that a base URL doesn't need to be
provided).