[css-navigation-1] Factor out current from/to URL definitions. (#13160)
This extracts the defintions of current from URL and current to URL from
the definition of the @navigation rule, which both allows them to be
used more formally in the definition of the ::link-to() pseudo-class and
also makes the definitions a bit more readable.

This also cleans up some whitespace in the spec.